restructured folder system and moved  components to features folder
