+++ client/src/features/auth/LoginForm.tsx
@@ -1,48 +1,42 @@
 import { Button } from "@/components/ui/button";
 import { Form } from "@/components/ui/form";
+import { useAuthService } from "@/services/authService";
 
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
-import { z } from "zod";
-
 import { useNavigate } from "react-router-dom";
-import { useAuthService } from "@/services/authService";
-import { FormFieldWrapper } from "./FormFieldWrapper";
+import { z } from "zod";
+import { FormFieldWrapper } from "../../components/FormFieldWrapper";
 
 const formSchema = z.object({
   email: z.string().email(),
   password: z
     .string()
     .min(6, { message: "Password must contains at least 6 character" }),
-  firstName: z.string().min(1, { message: "First  name is required" }),
-  lastName: z.string().min(1, { message: "Last name is required" }),
 });
 
-export default function RegisterForm() {
-  const { register, isLoading, error } = useAuthService();
-
+export default function LoginForm() {
   const navigate = useNavigate();
+  const { login, isLoading, error } = useAuthService();
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
       email: "",
       password: "",
-      firstName: "",
-      lastName: "",
     },
   });
 
   async function onSubmit(values: z.infer<typeof formSchema>) {
-    const result = await register(values);
+    const result = await login(values);
 
     if (result) {
-      console.log(result);
-      navigate("/company");
+      navigate("/dashboard");
     } else {
       console.log(error);
     }
   }
+
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(onSubmit)}>
@@ -50,7 +44,7 @@ export default function RegisterForm() {
           name="email"
           control={form.control}
           label="Email"
-          placeholder="Enter your company name..."
+          placeholder="Enter your email..."
         />
         <FormFieldWrapper
           name="password"
@@ -59,22 +53,10 @@ export default function RegisterForm() {
           placeholder="Enter your password..."
           type="password"
         />
-        <FormFieldWrapper
-          name="firstName"
-          control={form.control}
-          label="First name"
-          placeholder="Enter your first name..."
-        />
-        <FormFieldWrapper
-          name="lastName"
-          control={form.control}
-          label="Last name"
-          placeholder="Enter your last name..."
-        />
-        {!isLoading && error && <p className="text-red-500">{error}</p>}
+        {error && !isLoading && <p className="text-red-500">{error}</p>}
         <div className="mt-4 w-full">
           <Button className="w-full" type="submit" disabled={isLoading}>
-            {isLoading ? "Registering..." : "Register"}
+            {isLoading ? "Logging in..." : "Login"}
           </Button>
         </div>
       </form>
